feat: Move presets/labelBehavior to sub-packages
