+++ packages/react-responsive-pagination/labelBehaviour.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/labelBehaviour.js';
