+++ packages/react-responsive-pagination/labelBehaviour.js
@@ -1,0 +1,1 @@
+export * from './dist/labelBehaviour.js';
