+++ packages/react-responsive-pagination/presets.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/presets.js';
