+++ packages/react-responsive-pagination/presets.js
@@ -1,0 +1,1 @@
+export * from './dist/presets.js';
