+++ packages/react-responsive-pagination/src/presets.tsx
@@ -1,0 +1,10 @@
+import { srOnlySpanLabel } from './labelBehaviour.js';
+
+export const v1_bootstrap4PaginationPreset = {
+  ariaCurrentAttr: false,
+  labelBehaviour: srOnlySpanLabel(),
+};
+
+export const bootstrap4PaginationPreset = {};
+
+export const bootstrap5PaginationPreset = {};
